corrosion::persistent

Module client

Code for clients that connect to a remote crate::persistent::server::Server and either send mutations, or subscribe to said mutations

Structs

Client

A persistent connection to a corrosion server

ExecResponse

MutationClient

SubClientStream

SubscriptionClient

Enums

ConnectError

ExecResult

StreamError

SubscriptionStop

TransactionError

Type Aliases

QueryEvent