corrosion

Module pubsub

Reimplementation of some pieces of https://github.com/superfly/corrosion/blob/main/crates/corro-agent/src/api/public/pubsub.rs

The Corrosion code was too linked to HTTP, which we may or may not use

Structs

BufferingSubStream

An async stream that buffers events, used by senders to batch changes

ChangeId

ChangeId newtype to differentiate from RowId

MatcherCache

MatcherHandle

PubsubContext

The context needed to create and manage subscriptions

SubParamsv1

SubsManager

Subscription

SubscriptionEvent

SubscriptionStream

The read side of a BufferingSubStream

Enums

CatchUpError

MatcherError

MatcherUpsertError

TypedQueryEvent

Constants

SERVER_QUERY

Functions

catch_up_sub

Given a new subscriber to an existing matcher, attempts to catch up the subscriber to the current state of the database

catch_up_sub_anew

catch_up_sub_from

change_threshold

The default [SubParams::change_threshold]

error_to_sub_event

max_time

The default [SubParams::max_time]

process_interval

The default [SubParams::process_interval]

process_sub_channel

Processes a single matcher mpsc channel to sent it to the broadcast channel in the case of multiple subscriptions to the same query

query_to_sub_event

read_length_prefixed_bytes

Reads the next length-prefixed chunk from the bytes

restore_subscriptions

Initialize subscription state and tasks

subscribe

Creates a subscription for the specified [PubusbContext]

upsert_sub

Upserts a subscription

Type Aliases

BodySender

MatchCandidates

MatcherCacheInner

QueryEvent

SharedMatcherCache